requirements.txt
# General packages
scipy==1.6.1
networkx>=2.1
numpy==1.20.1
pandas==1.2.2
scikit_learn==0.24.1
tqdm==4.57.0
PyYAML==5.4.1
jupyter==1.0.0
seaborn==0.11.1
matplotlib==3.3.1
optuna==2.10.1
wandb==0.13.2
plotly==5.10.0
typer[all]==0.6.1
darker==1.5.0

# DVC
dvc[s3]==2.1.0
pathspec==0.9.0

# PyTorch Geometric
--find-links https://download.pytorch.org/whl/torch_stable.html
--find-links https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html

torch==1.7.0+cu110
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-cluster==1.5.8
torch-spline-conv==1.2.0
torch-geometric==1.7.0

# Graph Embedding library
# (The official implementation uses a Gensim version that is
# incompatible with karateclub, but the two models (LINE, SDNE)
# we utilize here, do not utilize Gensim at all. I updated the
# dependency versions so it can be installed without additional hacks)
git+https://github.com/pbielak/GraphEmbedding

# Open Graph Benchmark
ogb>=1.3.0

# KarateClub
karateclub==1.0.24
